move reducers into reducer files within dirs
