+++ src/store/quote/reducer.js
